requirements.txt
fastapi
uvicorn
python-multipart
boto3
pydantic
python-dotenv
confluent-kafka
pyyaml



torch 
torchvision 
transformers 
opencv-python 
pillow
doctr[torch] 
python-dotenv
accelerate